+++ src/components/Play.js
@@ -12,6 +12,7 @@ import api from '../api';
 import IdLink from './IdLink';
 import Years from './Years';
 import CastList from './CastList';
+import DownloadLinks from './DownloadLinks';
 import NetworkGraph from './NetworkGraph';
 import RelationsGraph from './RelationsGraph';
 import SpeechDistribution from './SpeechDistribution';
@@ -24,7 +25,7 @@ const apiUrl = api.getBaseURL();
 const edgeColor = '#61affe65';
 const nodeColor = '#61affe';
 
-const tabNames = ['network', 'speech', 'text', 'relations'];
+const tabNames = ['network', 'speech', 'text', 'relations', 'downloads'];
 
 function getCooccurrences (segments) {
   const map = {};
@@ -90,7 +91,6 @@ const PlayInfo = ({corpusId, playId}) => {
   const [play, setPlay] = useState(null);
   const [graph, setGraph] = useState(null);
   const [error, setError] = useState(null);
-  const [box, setBox] = useState('cast');
 
   useEffect(() => {
     async function fetchPlay () {
@@ -116,11 +116,6 @@ const PlayInfo = ({corpusId, playId}) => {
 
     fetchPlay();
   }, [corpusId, playId]);
-
-  function toggle (b) {
-    setBox(b === box ? null : b);
-    console.log(box);
-  }
 
   if (error && error.message === 'not found') {
     return <p>No such play!</p>;
@@ -152,17 +147,13 @@ const PlayInfo = ({corpusId, playId}) => {
     tab = 'network';
   }
 
-  const playUrl = `${apiUrl}/corpora/${play.corpus}/play/${play.name}`;
-  const csvUrl = `${playUrl}/networkdata/csv`;
-  const gexfUrl = `${playUrl}/networkdata/gexf`;
-  const graphmlUrl = `${playUrl}/networkdata/graphml`;
-  const csvRelationsUrl = `${playUrl}/relations/csv`;
-  const gexfRelationsUrl = `${playUrl}/relations/gexf`;
-  const teiUrl = `${playUrl}/tei`;
+  const teiUrl = `${apiUrl}/corpora/${play.corpus}/play/${play.name}/tei`;
 
   let tabContent = null;
   if (tab === 'speech') {
     tabContent = <SpeechDistribution segments={play.segments} {...{groups}}/>;
+  } else if (tab === 'downloads') {
+    tabContent = <DownloadLinks play={play}/>;
   } else if (tab === 'text') {
     tabContent = <TEIPanel url={teiUrl}/>;
   } else if (tab === 'relations') {
@@ -222,103 +213,6 @@ const PlayInfo = ({corpusId, playId}) => {
           <li>
             DraCor: <a href={`/id/${play.id}`}>{play.id}</a>
           </li>
-          <li className="play-downloads">
-            Downloads:{' '}
-            <span
-              className="play-downloads-item"
-              onClick={() => toggle('nd-formats')}
-            >
-              <i>network&nbsp;data</i>
-              {box === 'nd-formats' && (
-                <span className="formats">
-                  <a href={csvUrl} download={`${play.id}-${play.name}.csv`}>
-                    CSV
-                  </a>
-                  <a href={gexfUrl} download={`${play.id}-${play.name}.gexf`}>
-                    GEXF
-                  </a>
-                  <a
-                    href={graphmlUrl}
-                    download={`${play.id}-${play.name}.graphml`}
-                  >
-                    GraphML
-                  </a>
-                </span>
-              )}
-            </span>{' '}
-            <span
-              className="play-downloads-item"
-              onClick={() => toggle('rd-formats')}
-            >
-              <i>relation&nbsp;data</i>
-              {box === 'rd-formats' && (
-                <span className="formats">
-                  <a
-                    href={csvRelationsUrl}
-                    download={`${play.id}-${play.name}-relations.csv`}
-                  >
-                    CSV
-                  </a>
-                  <a
-                    href={gexfRelationsUrl}
-                    download={`${play.id}-${play.name}-relations.gexf`}
-                  >
-                    GEXF
-                  </a>
-                </span>
-              )}
-            </span>{' '}
-            <span
-              className="play-downloads-item"
-              onClick={() => toggle('st-formats')}
-            >
-              <i>spoken&nbsp;text</i>
-              {box === 'st-formats' && (
-                <span className="formats">
-                  <a
-                    href={`${playUrl}/spoken-text`}
-                    download={`${play.id}-${play.name}-spoken.txt`}
-                  >
-                    TXT
-                  </a>
-                </span>
-              )}
-            </span>{' '}
-            <span
-              className="play-downloads-item"
-              onClick={() => toggle('stc-formats')}
-            >
-              <i>spoken&nbsp;text&nbsp;by&nbsp;character</i>
-              {box === 'stc-formats' && (
-                <span className="formats">
-                  <a
-                    href={`${playUrl}/spoken-text-by-character.json`}
-                    download={`${play.id}-${play.name}-spoken.json`}
-                  >
-                    JSON
-                  </a>
-                </span>
-              )}
-            </span>{' '}
-            <span
-              className="play-downloads-item"
-              onClick={() => toggle('sd-formats')}
-            >
-              <i>stage&nbsp;directions</i>
-              {box === 'sd-formats' && (
-                <span className="formats">
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href={`${playUrl}/stage-directions`}
-                    download={`${play.id}-${play.name}-stage.txt`}
-                  >
-                    TXT
-                  </a>
-                </span>
-              )}
-            </span>
-          </li>
         </ul>
       </hgroup>
 
@@ -373,6 +267,14 @@ const PlayInfo = ({corpusId, playId}) => {
               className={classnames({active: tab === 'text'})}
             >
               Full text
+            </NavLink>
+          </NavItem>
+          <NavItem>
+            <NavLink
+              href="#downloads"
+              className={classnames({active: tab === 'downloads'})}
+            >
+              Downloads
             </NavLink>
           </NavItem>
         </Nav>
